agents/compare_response/pyproject.toml
[project]
name = "compare-response"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.3.0",
    "langchain-openai>=1.1.1",
    "langgraph>=1.0.4",
    "strands-agents>=1.19.0",
]
